packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-memory"
version = "0.1.0"
description = "Persistent cross-session memory, recalled facts, and agent knowledge base"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["memory","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-memory"
Repository = "https://github.com/mxn2020/minions-memory"

[tool.hatch.build.targets.wheel]
packages = ["minions_memory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
